Extract parsing of rules
