+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/visualizers/rules/CardinalityRuleVisualizer.java
@@ -3,24 +3,35 @@ package org.archcnl.domain.input.visualization.visualizers.rules;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.regex.Pattern;
+import java.util.Set;
 import org.archcnl.domain.common.ConceptManager;
 import org.archcnl.domain.common.RelationManager;
-import org.archcnl.domain.input.model.architecturerules.ArchitectureRule;
+import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Triplet;
+import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Variable;
 import org.archcnl.domain.input.visualization.coloredmodel.ColorState;
 import org.archcnl.domain.input.visualization.exceptions.MappingToUmlTranslationFailedException;
+import org.archcnl.domain.input.visualization.visualizers.rules.rulemodel.RuleVariant;
 import org.archcnl.domain.input.visualization.visualizers.rules.rulemodel.VerbPhrase;
+import org.archcnl.domain.input.visualization.visualizers.rules.rulemodel.VerbPhraseContainer;
 
 public class CardinalityRuleVisualizer extends RuleVisualizer {
 
-    private static final Pattern CNL_PATTERN =
-            Pattern.compile(
-                    "Every " + "(a |an )?" + SUBJECT_REGEX + " can " + PHRASES_REGEX + "\\.");
-
     public CardinalityRuleVisualizer(
-            ArchitectureRule rule, ConceptManager conceptManager, RelationManager relationManager)
+            String cnlString,
+            List<Triplet> subjectTriplets,
+            VerbPhraseContainer verbPhrases,
+            ConceptManager conceptManager,
+            RelationManager relationManager,
+            Set<Variable> usedVariables)
             throws MappingToUmlTranslationFailedException {
-        super(rule, conceptManager, relationManager);
+        super(
+                cnlString,
+                subjectTriplets,
+                verbPhrases,
+                conceptManager,
+                relationManager,
+                usedVariables);
+        transformToDiagram();
     }
 
     @Override
@@ -77,14 +88,5 @@ public class CardinalityRuleVisualizer extends RuleVisualizer {
             wrongPostfix.append("W");
         }
         return variants;
-    }
-
-    public static boolean matches(ArchitectureRule rule) {
-        return CNL_PATTERN.matcher(rule.toString()).matches();
-    }
-
-    @Override
-    protected Pattern getCnlPattern() {
-        return CNL_PATTERN;
     }
 }
