+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/visualizers/rules/NegationRuleVisualizer.java
@@ -3,49 +3,37 @@ package org.archcnl.domain.input.visualization.visualizers.rules;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+import java.util.Set;
 import java.util.stream.Collectors;
 import org.archcnl.domain.common.ConceptManager;
 import org.archcnl.domain.common.RelationManager;
+import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Triplet;
 import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Variable;
-import org.archcnl.domain.input.model.architecturerules.ArchitectureRule;
 import org.archcnl.domain.input.visualization.coloredmodel.ColorState;
 import org.archcnl.domain.input.visualization.coloredmodel.ColoredTriplet;
 import org.archcnl.domain.input.visualization.exceptions.MappingToUmlTranslationFailedException;
+import org.archcnl.domain.input.visualization.visualizers.rules.rulemodel.RuleVariant;
 import org.archcnl.domain.input.visualization.visualizers.rules.rulemodel.VerbPhrase;
+import org.archcnl.domain.input.visualization.visualizers.rules.rulemodel.VerbPhraseContainer;
 
 public class NegationRuleVisualizer extends RuleVisualizer {
 
-    private static final Pattern CNL_PATTERN =
-            Pattern.compile(
-                    "(Nothing|No (a |an )?"
-                            + SUBJECT_REGEX
-                            + ")"
-                            + " can "
-                            + PHRASES_REGEX
-                            + "\\.");
-
     public NegationRuleVisualizer(
-            ArchitectureRule rule, ConceptManager conceptManager, RelationManager relationManager)
+            String cnlString,
+            List<Triplet> subjectTriplets,
+            VerbPhraseContainer verbPhrases,
+            ConceptManager conceptManager,
+            RelationManager relationManager,
+            Set<Variable> usedVariables)
             throws MappingToUmlTranslationFailedException {
-        super(rule, conceptManager, relationManager);
-    }
-
-    @Override
-    protected void parseRule(String ruleString) throws MappingToUmlTranslationFailedException {
-        Matcher matcher = getCnlPattern().matcher(ruleString);
-        Helper.tryToFindMatch(matcher);
-        String phrasesGroup = matcher.group("phrases");
-        verbPhrases = parseVerbPhrases(phrasesGroup);
-        if (isNothingRule()) {
-            var nothingVar = new Variable("nothing");
-            var relation = verbPhrases.getPhrases().get(0).getPredicate().getRelation();
-            var triplet = Helper.getBaseSubjectTypeTriplet(relation, nothingVar, usedVariables);
-            subjectTriplets = Arrays.asList(triplet);
-        } else {
-            subjectTriplets = parseConceptExpression(matcher.group("subject"));
-        }
+        super(
+                cnlString,
+                subjectTriplets,
+                verbPhrases,
+                conceptManager,
+                relationManager,
+                usedVariables);
+        transformToDiagram();
     }
 
     @Override
@@ -84,18 +72,5 @@ public class NegationRuleVisualizer extends RuleVisualizer {
             postfix.append("1");
         }
         return variants;
-    }
-
-    public static boolean matches(ArchitectureRule rule) {
-        return CNL_PATTERN.matcher(rule.toString()).matches();
-    }
-
-    private boolean isNothingRule() {
-        return cnlString.startsWith("Nothing");
-    }
-
-    @Override
-    protected Pattern getCnlPattern() {
-        return CNL_PATTERN;
     }
 }
