+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/visualizers/rules/SubconceptRuleVisualizer.java
@@ -2,46 +2,35 @@ package org.archcnl.domain.input.visualization.visualizers.rules;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+import java.util.Set;
 import org.archcnl.domain.common.ConceptManager;
 import org.archcnl.domain.common.RelationManager;
-import org.archcnl.domain.input.model.architecturerules.ArchitectureRule;
+import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Triplet;
+import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Variable;
 import org.archcnl.domain.input.visualization.coloredmodel.ColorState;
 import org.archcnl.domain.input.visualization.exceptions.MappingToUmlTranslationFailedException;
-import org.archcnl.domain.input.visualization.helpers.IsARelation;
-import org.archcnl.domain.input.visualization.visualizers.rules.rulemodel.RulePredicate;
+import org.archcnl.domain.input.visualization.visualizers.rules.rulemodel.RuleVariant;
 import org.archcnl.domain.input.visualization.visualizers.rules.rulemodel.VerbPhrase;
 import org.archcnl.domain.input.visualization.visualizers.rules.rulemodel.VerbPhraseContainer;
 
 public class SubconceptRuleVisualizer extends RuleVisualizer {
 
-    private static final Pattern CNL_PATTERN =
-            Pattern.compile(
-                    "Every "
-                            + "(a |an )?"
-                            + SUBJECT_REGEX
-                            + " must be"
-                            + " (a |an )?"
-                            + OBJECT_REGEX
-                            + "\\.");
-
     public SubconceptRuleVisualizer(
-            ArchitectureRule rule, ConceptManager conceptManager, RelationManager relationManager)
+            String cnlString,
+            List<Triplet> subjectTriplets,
+            VerbPhraseContainer verbPhrases,
+            ConceptManager conceptManager,
+            RelationManager relationManager,
+            Set<Variable> usedVariables)
             throws MappingToUmlTranslationFailedException {
-        super(rule, conceptManager, relationManager);
-    }
-
-    @Override
-    protected void parseRule(String ruleString) throws MappingToUmlTranslationFailedException {
-        Matcher matcher = getCnlPattern().matcher(ruleString);
-        Helper.tryToFindMatch(matcher);
-        var isAPredicate = new RulePredicate(new IsARelation());
-        var objectTriplets = parseConceptExpression(matcher.group("object"));
-        VerbPhrase phrase = new VerbPhrase(isAPredicate, objectTriplets);
-        verbPhrases = new VerbPhraseContainer();
-        verbPhrases.addVerbPhrase(phrase);
-        subjectTriplets = parseConceptExpression(matcher.group("subject"));
+        super(
+                cnlString,
+                subjectTriplets,
+                verbPhrases,
+                conceptManager,
+                relationManager,
+                usedVariables);
+        transformToDiagram();
     }
 
     @Override
@@ -66,14 +55,5 @@ public class SubconceptRuleVisualizer extends RuleVisualizer {
     protected List<RuleVariant> buildRuleVariantsOr()
             throws MappingToUmlTranslationFailedException {
         throw new UnsupportedOperationException("The sub-concept rule type doesn't allow OR");
-    }
-
-    public static boolean matches(ArchitectureRule rule) {
-        return CNL_PATTERN.matcher(rule.toString()).matches();
-    }
-
-    @Override
-    protected Pattern getCnlPattern() {
-        return CNL_PATTERN;
     }
 }
