+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/visualizers/rules/rulemodel/RuleVariant.java
@@ -1,4 +1,4 @@
-package org.archcnl.domain.input.visualization.visualizers.rules;
+package org.archcnl.domain.input.visualization.visualizers.rules.rulemodel;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,7 +6,6 @@ import java.util.Optional;
 import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Variable;
 import org.archcnl.domain.input.visualization.coloredmodel.ColorState;
 import org.archcnl.domain.input.visualization.coloredmodel.ColoredTriplet;
-import org.archcnl.domain.input.visualization.visualizers.rules.rulemodel.RulePredicate;
 
 public class RuleVariant {
